- continued new implementation for UWP+DX12

+++ src/rendersystems/DirectX/12/default-shaders/PixelShader_TexturedMultiply.hlsl
@@ -1,8 +1,8 @@
-/// @version 4.1
+/// @version 5.1
 
 #include "PixelShader_Textured.hlsli"
 
-min16float4 main(PixelShaderInput input) : SV_Target
+float4 main(PixelShaderInput input) : SV_Target
 {
 	return (cTexture.Sample(cSampler, input.tex) * input.color);
 }
