+++ src/rendersystems/DirectX/12/default-shaders/VertexShader_ColoredTextured.hlsl
@@ -1,19 +1,19 @@
-/// @version 4.1
+/// @version 5.1
 
 #include "VertexShader.hlsli"
 #include "PixelShader_Textured.hlsli"
 
 struct VertexShaderInput
 {
-	min16float3 position : POSITION;
-	min16float4 color : COLOR;
+	float3 position : POSITION;
+	float4 color : COLOR;
 	float2 tex : TEXCOORD0;
 };
 
 PixelShaderInput main(VertexShaderInput input)
 {
 	PixelShaderInput vertexShaderOutput;
-	vertexShaderOutput.position = mul(min16float4(input.position, (min16float)1.0), cMatrix);
+	vertexShaderOutput.position = mul(float4(input.position, 1.0), cMatrix);
 	vertexShaderOutput.color = input.color;
 	vertexShaderOutput.tex = input.tex;
 	vertexShaderOutput.lerpAlpha = cLerpAlpha;
