+++ src/rendersystems/DirectX/12/default-shaders/VertexShader_Textured.hlsl
@@ -1,18 +1,18 @@
-/// @version 4.1
+/// @version 5.1
 
 #include "VertexShader.hlsli"
 #include "PixelShader_Textured.hlsli"
 
 struct VertexShaderInput
 {
-	min16float3 position : POSITION;
+	float3 position : POSITION;
 	float2 tex : TEXCOORD0;
 };
 
 PixelShaderInput main(VertexShaderInput input)
 {
 	PixelShaderInput vertexShaderOutput;
-	vertexShaderOutput.position = mul(min16float4(input.position, (min16float)1.0), cMatrix);
+	vertexShaderOutput.position = mul(float4(input.position, 1.0), cMatrix);
 	vertexShaderOutput.color = cSystemColor;
 	vertexShaderOutput.tex = input.tex;
 	vertexShaderOutput.lerpAlpha = cLerpAlpha;
